requirements.txt
# Python requirements for Multi-Agent System

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Redis
redis>=5.0.0
hiredis>=2.2.0

# GitHub integration
PyGithub>=2.1.0

# HTTP client
aiohttp>=3.9.0
httpx>=0.25.0

# LLM integrations - LangChain + LangGraph (compatible versions)
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.5,<0.1.0
langchain-anthropic>=0.1.0,<0.2.0
langchain-community>=0.0.13,<0.1.0
langgraph>=0.0.20,<0.1.0

# Vector stores for RAG
chromadb>=0.4.22,<0.5.0
faiss-cpu>=1.8.0
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
pyyaml>=6.0.0
click>=8.1.0

# Logging and monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.13.0
